+++ crates/aos-kernel/src/capability.rs
@@ -2,35 +2,12 @@ use std::collections::HashMap;
 
 use aos_air_types::{CapGrant, CapGrantBudget, CapType, DefCap, Manifest, Name, ValueLiteral};
 use aos_cbor::to_canonical_cbor;
-use aos_effects::traits::PolicyDecision;
-use aos_effects::{CapabilityBudget, CapabilityGrant, EffectIntent, EffectSource};
+use aos_effects::{CapabilityBudget, CapabilityGrant};
 
 use crate::error::KernelError;
 
 pub trait CapabilityGate {
     fn resolve(&self, cap_name: &str, effect_kind: &str) -> Result<CapabilityGrant, KernelError>;
-}
-
-pub trait PolicyGate {
-    fn decide(
-        &self,
-        intent: &EffectIntent,
-        grant: &CapabilityGrant,
-        source: &EffectSource,
-    ) -> Result<PolicyDecision, KernelError>;
-}
-
-pub struct AllowAllPolicy;
-
-impl PolicyGate for AllowAllPolicy {
-    fn decide(
-        &self,
-        _intent: &EffectIntent,
-        _grant: &CapabilityGrant,
-        _source: &EffectSource,
-    ) -> Result<PolicyDecision, KernelError> {
-        Ok(PolicyDecision::Allow)
-    }
 }
 
 #[derive(Clone)]
